pyproject.toml
[project]
name = "django-brazilian-zipcode"
version = "0.1.0"
description = ""
authors = [
    {name = "Plathanus", email = "leandrodesouzadev@gmail.com"},
]
dependencies = [
    "requests>=2.27.1",
    "django==3.2",
    "djangorestframework>=3.13.1",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[project.urls]
Homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.0.1",
    "pytest-mock>=3.7.0",
    "tox>=3.24.5",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
